+++ api/src/formatters/jsonnested.ts
@@ -1,3 +1,4 @@
+import { config } from '../config';
 import { Exporter, IntermediateTranslation, IntermediateTranslationFormat, Parser } from '../domain/formatters';
 
 export const JSON_MAX_NESTED_LEVELS = 100;
@@ -11,8 +12,8 @@ export const jsonNestedParser: Parser = async (data: string) => {
   }
 
   const traverse = (obj, level = 0, parentTerm = undefined) => {
-    if (level > JSON_MAX_NESTED_LEVELS) {
-      throw new Error(`Too many nested levels in JSON content (>${JSON_MAX_NESTED_LEVELS})`);
+    if (level > config.import.maxNestedLevels) {
+      throw new Error(`Too many nested levels in JSON content (>${config.import.maxNestedLevels})`);
     }
     for (const key of Object.keys(obj)) {
       const value = obj[key];